+++ src/main/java/edu/blakealmanza/controller/CustomerController.java
@@ -1,0 +1,5 @@
+package edu.blakealmanza.controller;
+
+public class CustomerController {
+
+}
