--- src/main/java/edu/blakealmanza/controllers/WebController.java
@@ -1,21 +1,0 @@
-package edu.blakealmanza.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class WebController {
-
-  @GetMapping("/")
-  public String home(Model model) {
-    model.addAttribute("name", "Blake");
-    return "index";
-  }
-
-  @GetMapping("/products")
-  public String showAllProducts(Model model) {
-    return "products";
-  }
-
-}
