+++ src/main/java/edu/blakealmanza/model/Order.java
@@ -1,0 +1,5 @@
+package edu.blakealmanza.model;
+
+public class Order {
+
+}
