+++ src/main/resources/templates/manager/products.html
@@ -16,7 +16,7 @@
   <a href="/">Back to Home</a>
 
   <!-- Load the JS file here -->
-  <script src="/js/products.js"></script>
+  <script src="/js/manager.js"></script>
 </body>
 
 </html>
